+++ lib/tableficate/exceptions.rb
@@ -1,3 +1,7 @@
 module Tableficate
   class MissingScope < StandardError; end
+  
+  module Filter
+    class MissingTemplate < StandardError; end
+  end
 end