+++ lib/tableficate/filters/input_field.rb
@@ -1,0 +1,6 @@
+module Tableficate
+  module Filter
+    class InputField < Base
+    end
+  end
+end