+++ lib/tableficate/filters/select.rb
@@ -1,0 +1,6 @@
+module Tableficate
+  module Filter
+    class Select < Base
+    end
+  end
+end