+++ lib/tableficate/filters/select_range.rb
@@ -1,0 +1,6 @@
+module Tableficate
+  module Filter
+    class SelectRange < Base
+    end
+  end
+end